+++ engine/input/maps/InputMap.cpp
@@ -70,7 +70,7 @@ namespace input {
             return m_analogMap.empty() && m_analogStateMap.empty() && m_keyMap.empty();
         }
 
-        void InputMap::setValue(const engine::input::AnalogKeys key, const int value)
+        void InputMap::setValue(const engine::input::AnalogKeys key, const double value)
         {
             if (key == AnalogKeys::CON_UNKNOWN) {
                 return;
@@ -102,7 +102,7 @@ namespace input {
             return m_analogStateMap.at(key);
         }
 
-        int InputMap::getValue(const AnalogKeys key) const
+        double InputMap::getValue(const AnalogKeys key) const
         {
             if (m_analogMap.find(key) == m_analogMap.end()) {
                 return 0;
